# swagger.yaml
openapi: 3.0.3
info:
  title: SLSEA Solar Generation Data API
  description: >
    Real-time and historical solar generation data API for the Sri Lanka
    Sustainable Energy Authority. Metering devices push readings as write
    clients; SLSEA users read data scoped by jurisdiction.
  version: 1.0.0
servers:
  - url: https://webapicw-production.up.railway.app
    description: Production (Railway)
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /substations/{id}:
    get:
      tags: [Substations]
      summary: Get a single substation (atomic resource)
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: One substation (with populated district and province)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Substation'
        '404':
          $ref: '#/components/responses/NotFound'
  /substations/{id}/installations:
    get:
      tags: [Substations]
      summary: List installations connected to this substation
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Array of installations at this substation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstallationBasic'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /installations/{id}/readings:
    get:
      tags: [Readings]
      summary: Paginated, filterable, sortable reading history (analytical resource)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - name: from
          in: query
          description: ISO-8601 date or date-time — start of time window filter
          schema: { type: string, format: date }
        - name: to
          in: query
          description: ISO-8601 date or date-time — end of time window filter (must not be earlier than `from`)
          schema: { type: string, format: date }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
      responses:
        '200':
          description: Paginated readings with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reading'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Installation is outside your jurisdiction
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Readings]
      summary: Device pushes a new generation reading (write path, JWT-authenticated)
      description: >
        Requires a DEVICE token (installation-write scope) bound to this
        installation. SLSEA user tokens are rejected with 403 WRONG_TOKEN_TYPE.
        The installation must still exist (404 otherwise).
      security:
        - deviceBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [powerKw, energyKwh, voltage]
              properties:
                timestamp:
                  type: string
                  format: date-time
                  description: ISO-8601. Defaults to the server time. May not be more than 5 minutes in the future.
                powerKw: { type: number, minimum: 0 }
                energyKwh: { type: number, minimum: 0, description: Cumulative energy meter value }
                voltage: { type: number, minimum: 0 }
      responses:
        '201':
          description: Reading created
          headers:
            Location:
              description: URL of the newly created reading
              schema: { type: string }
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: >
            WRONG_TOKEN_TYPE (a user token was used), WRONG_SCOPE, or
            WRONG_INSTALLATION (token is bound to a different installation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
  /installations/{id}/readings/{readingId}:
    get:
      tags: [Readings]
      summary: Get a single reading (atomic resource)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: readingId
          in: path
          required: true
          schema: { type: string }
          description: MongoDB ObjectId of the reading
      responses:
        '200':
          description: One reading
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reading'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Installation is outside your jurisdiction
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        SLSEA USER token from POST /auth/login (valid 8 hours). Used for every
        read/admin endpoint. A device token is rejected here with
        403 WRONG_TOKEN_TYPE, even though it is validly signed.
    deviceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        DEVICE token from POST /installations/{id}/token (valid 1 hour, scope
        installation-write, bound to one installation). Used only to push
        readings. A user token is rejected here with 403 WRONG_TOKEN_TYPE.
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Device API key, issued at installation provisioning
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: '^[a-fA-F0-9]{24}$' }
      description: MongoDB ObjectId (24 hex characters). A malformed value is a 400 INVALID_ID.
  responses:
    ValidationError:
      description: >
        Malformed id (INVALID_ID), invalid JSON (INVALID_JSON) or a field that failed
        validation (VALIDATION_ERROR). For VALIDATION_ERROR and INVALID_ID, error.detail
        is a list of `{ location, field, message }` entries — one per problem.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    PayloadTooLarge:
      description: Request body exceeds the size limit (PAYLOAD_TOO_LARGE)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnsupportedMediaType:
      description: Request body was not sent as application/json (UNSUPPORTED_MEDIA_TYPE)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Substation:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
        district:
          type: object
          properties:
            _id: { type: string }
            name: { type: string }
            code: { type: string }
            province:
              type: object
              properties:
                _id: { type: string }
                name: { type: string }
                code: { type: string }
...
    InstallationBasic:
      type: object
      description: Shape returned by GET /substations/{id}/installations — substation is just an ObjectId reference (already known from the URL)
      properties:
        _id: { type: string }
        meterId: { type: string }
        capacityKw: { type: number }
        latitude: { type: number }
        longitude: { type: number }
        substation: { type: string, description: "ObjectId reference to this substation" }
    Reading:
      type: object
      properties:
        _id: { type: string }
        installation: { type: string }
        timestamp: { type: string, format: date-time }
        powerKw: { type: number }
        energyKwh: { type: number }
        voltage: { type: number }
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
        nextPage: { type: string, nullable: true }
        prevPage: { type: string, nullable: true }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            detail: { nullable: true }